[Sonar] More...
Sonar doesn't reliably report issues across the whole project when
"Show Active File Only" is disabled. I enabled it and went through
each file individually. There are still a few nitpicks but I'm leaving them
in without disabling the corresponding rules.

+++ src/com/characterforming/jrte/engine/ModelDecompiler.java
@@ -50,7 +50,7 @@ public class ModelDecompiler {
 		Set<Map.Entry<Bytes, Integer>> effectorOrdinalMap = this.model.getEffectorOrdinalMap().entrySet();
 		String[] effectorNames = new String[effectorOrdinalMap.size()];
 		for (Map.Entry<Bytes, Integer> entry : effectorOrdinalMap) {
-			effectorNames[(int)entry.getValue()] = Bytes.decode(this.decoder, entry.getKey().getData(), entry.getKey().getLength()).toString();
+			effectorNames[entry.getValue()] = Bytes.decode(this.decoder, entry.getKey().getData(), entry.getKey().getLength()).toString();
 		}
 		System.err.printf("%s%n%nInput equivalents (equivalent: input...)%n%n", transducerName);
 		for (int i = 0; i < inputEquivalentCount; i++) {
